Twitter effect on Covid
The global health care infrastructure, as well as the social, economic, and psychological well-being of mankind, are suffering as a result of the current coronavirus illness (COVID-19) pandemic that began in late 2018. The main objective of this study was to identify the relationship between COVID-19 Twitter hashtags and the daily increase in COVID-19 cases. The study used daily time-series data for 12 months, from 2020 March to 2021 February. This data was obtained from Worldometers and GitHub (Lopez Bec) websites, bulletins, and other publications. Data was analyzed and analyzed by the use of correlation analysis as well as bivariate regression analysis. The finding of the study revealed that there existed a positive association between daily, weekly, and monthly increases in COVID-19 and those of Twitter hashtags, also based on daily, weekly, and monthly datasets. Though a strong and positive association for monthly data existed between the two datasets. Additionally, the regression analysis revealed that a positive and statistically significant relationship between Twitter hashtags and the daily increase of COVID-19 existed. Finally, the research noticed that on-the-ground and online public health crisis response operations are becoming more concurrent and linked. Social media platforms enable the direct communication of health information to the general audience. Through social media monitoring, health systems should aim to develop national and worldwide illness detection and surveillance systems.